requirements.txt

flask
transformers
torch
python-telegram-bot
requests
